requirements.txt
boto3
dnspython
fastapi
python-dotenv
python-multipart
uvicorn[standard]
requests
